requirements.txt
requests
pandas
airflow
psycopg2-binary